+++ check_display_others.go
@@ -1,0 +1,8 @@
+//go:build !linux && !darwin
+// +build !linux,!darwin
+
+package webbrowser
+
+func DisplayValid() (err error) {
+	return nil
+}
