+++ command_darwin.go
@@ -1,0 +1,7 @@
+//go:build darwin
+// +build darwin
+
+package webbrowser
+
+// command to execute when running on a mac.
+var openCommand *browserCommand = &browserCommand{"open", nil}
