+++ command_invalid.go
@@ -1,0 +1,6 @@
+//go:build !windows && !darwin && !android && !freebsd && !linux && !netbsd && !openbsd
+// +build !windows,!darwin,!android,!freebsd,!linux,!netbsd,!openbsd
+
+package webbrowser
+
+var openCommand *browserCommand = nil
